requirements.txt
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==4.5.4
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-GoogleMaps==0.2.6
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==latest
pandas==latest
price-parser==0.2.4
python-dotenv==0.10.3
requests==2.22.0
spacy-pytorch-transformers==latest
urllib3==1.25.6
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
